pyproject.toml
[project]
name = "book-llm-wiki"
version = "0.1.0"
description = "Two-tier book-to-wiki pipeline: local CLI converts EPUBs, Claude Code slash command summarizes into an Obsidian LLM wiki."
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
librarything = [
    "scrapling>=0.4",
]

[project.scripts]
book-llm-wiki = "book_llm_wiki.cli:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["book_llm_wiki*"]
